+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# Next
+
+- **[Fix]** Fix `Project::shared_with_groups` type.
+
 # 0.14.0
 
 - **[Breaking change]** Implement handlers as tower services.